Move navigation bar
In the near future, we will expand the site with links like
"Downloads" and "Wiki", and there is not enough space left to add them.

This commit moves the links below "Lineage OS Android Distribution",
and adds 2 drop down menus, "Contribute" and "People".

Change-Id: I88e6e0cd3551a637cdc6f11de89f3651c21a1fbb
Preview: http://imgur.com/a/BxO7s