Update patch set 2
Patch Set 2:

DO NOT MERGE YET. Breaks last week's post for some reason. Basically, on the home page, clicking last week's post leads to this one, probably because they share the same title.

Perhaps we should set the title (or permalink) to the week of the year? This post (w/c 13th Feb), is the 07th week

Patch-set: 2
Label: Verified=0