Update patch set 4
Patch Set 4: Code-Review+1

Sorry for the email spam. I've overrode the default "title=permalink" for this post, and set it manually to "/07-17-Last-week-in-LineageOS/", 07 being the 7th week of the year, and 17 being the year. 

I've confirmed this works on my machine. Be aware, the post won't show without changing its date, as it is dated for Monday. 

If anyone has a more elegant solution regarding permalinks, I'll be happy to hear it

Patch-set: 4
Label: Code-Review=+1