Update patch set 6
Patch Set 6:

Intro paragraph can change as this is now the second one of these:

"Welcome to LineageOS' week in review where we go over, foo".

The next header for "What's changed since then" needs a tie-back to the dates. Is it what's changed since the last post or since 2/13?

Security patch discussion needs an item for 13.0 branch: "We'll be working on incorporating the Feb security release into 13.0 builds in the near future."

Patch-set: 6
Label: Verified=0