Update patch set 12
Patch Set 12: Code-Review-1

Talked to brinly, there's no timeline on merging the multiple mcc stuff yet. If it makes it in before this changelog does, then it should be added. For now, i'm knocking it off, since I don't know when that's happening. Same deal with the EU numbers, I'm _more_ certain that that's happening soon, so I'm leaving it on for now. -1 as a reminder to myself that the situation on those has to be monitored

Patch-set: 12
Label: Code-Review=-1